LayoutTests/platform/mac-wk1/http/tests/security/basic-auth-module-import-propagate-expected.txt

http://127.0.0.1:8000/security/resources/protected-script.php?testId=0 - didReceiveAuthenticationChallenge - Responding with testUser:testPassword
CONSOLE MESSAGE: Origin http://127.0.0.1:8000 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: Origin http://127.0.0.1:8000 is not allowed by Access-Control-Allow-Origin.
PASS did load script with ./protected-script.php?testId=0.

PASS did load script with http://127.0.0.1:8000/security/resources/protected-script.php?testId=1.

PASS did not load script with https://127.0.0.1:8443/security/resources/protected-script.php?testId=2.

PASS did not load script with https://localhost:8443/security/resources/protected-script.php?testId=3.

PASS successfullyParsed is true

TEST COMPLETE
